backend/requirements.txt
fastapi[standard]>=0.118.0
psycopg[binary]>=3.2.10
torch>=2.8.0
torchvision>=0.23.0
transformers>=4.57.0
pillow>=11.3.0
numpy>=2.3.3
python-multipart>=0.0.20
uvicorn>=0.30.0
